Set up training on generator models, various other features & organization
--Scripts--
`training.py` runs training. `run.py` runs the model on a single audio file.

--Dataset--
Dataset loader loads audio data, creates a downsampled version. Data
augmentation is performed: A random short portions of the audio is used, and
before filtering, low-pass filters of random type and parameters are applied.

--Models--
There are two models in `models.py`. The first is an Audio UNet, which I could
not get to work very well. It wouldn't do much other than flip the original
spectrogram repeatedly in the higher frequencies. The second is an
implementation of WaveNet, which uses dilated convolution layers, with summed
skip connections. This actually seems to work.

--Config--
Configuration is stored in a yaml file, and loaded in as a global variable.

--Transforms--
Transforms (resampling, spectrogram, etc.) are now accessed through static
class methods in the transforms file

--Resampling Tests--
Ran tests to evaluate the quality of down- & up-sampling.